requirements.txt
#pip requirements
markovify
lxml
discord.py
praw

